@@ -23,7 +23,7 @@
     "issues": "https://github.com/frozzare/wp-short-url/issues"
   },
   "require": {
-    "php": ">=5.3.0",
+    "php": ">=5.4.7",
     "composer/installers": "~1.0",
     "frozzare/wp-test-suite": "~1.0"
   }
